vinyl: factor out function to lookup key in page
This function is a part of the run iterator API so we can't use it in
a reader thread. Let's make it an independent helper. As a good side
effect, we can now reuse it in the slice stream implementation.

(cherry picked from commit ac8ce0233a00a8ff47ded858dc87ca7ccd2bd832)