sql: enable SOUNDEX sql builtin
Needed for #4182

@TarantoolBot document
Title: Introduce SOUNDEX sql function

The SOUNDEX function returns a 4-character code that represents
the sound of the words in the argument. The result can be
compared to the results of the SOUNDEX function of other strings.

The current SOUNDEX function supports only Latin strings.

@kostja: fix test tap count; remove optional invocation; remove
trailing spaces; fix alignment.