box: do not check state in case of reconnect
Test box/net.box.test.lua checks state of the connection in case
of an error. It should be 'error_reconnect'. But, in cases where
testing was performed on a slow computer or in the case of a very
large load, it is possible that the connection status may change
from the 'error_reconnect' state to another state. This led to the
failure of the test. Since this check is not the main purpose of
the test, it is better to simply delete the check.

Closes #4335

(cherry picked from commit 77051a11029560eee8ad0adf83c8ac92e64a7d12)