ddl: fix vylog vs xlog sync for multi-statement transactions
Vinyl writes a signature of each statement that created or dropped or
modified an LSM tree to vylog. This is needed to match xlog rows to
vylog records on recovery. So alter_space_commit() passes txn->signature
to index_commit_create(), index_commit_drop(), and index_commit_modify()
methods. If a transaction has multiple DDL statements, all of them will
have the same signature. Per se this is fine. The problem is that on
recovery we apply statements one-by-one, not transactionally, and so we
pass signatures assigned to individual statements, not to transactions.
As a result, while recovering a multi-statement DDL transaction vinyl
gets different signatures for them instead of those that were used when
those statements were committed, which apparently breaks its assumptions
and leads to a crash or even a data loss.

To fix this issue, let's, rather than using txn->signature, make
alter_space_commit() pass the signature of the actual statement that
committed a DDL operation, which is pretty easy to compute given the
number of statements in a transaction at the time when the statement was
committed.

Fixes commit f266559b167b ("ddl: allow to execute non-yielding DDL
statements in transactions").

Closes #4350
Follow-up #4083