box: generalize memtx_multikey_tree methods
Updated multikey memtx_tree methods a bit to make possible to
reuse them in case of functional indexes. The BPS_TREE_IDENTICAL
is only used in bps_tree_delete_identical (when BPS_TREE_NO_DEBUG
macro is defined) therefore it may be reused to delete an
identical entry in case of functional index in further patches.

Needed for #1260